rename files and move things around
