+++ crates/viewer/re_dataframe_ui/src/filters/mod.rs
@@ -1,6 +1,7 @@
 mod boolean;
-mod filter;
-mod filter_ui;
+mod column_filter;
+mod column_filter_ui;
+mod filter_trait;
 mod numerical;
 mod parse_timestamp;
 mod string;
@@ -8,6 +9,6 @@ mod timestamp;
 mod timestamp_formatted;
 
 pub use self::{
-    boolean::*, filter::*, filter_ui::*, numerical::*, parse_timestamp::*, string::*, timestamp::*,
-    timestamp_formatted::*,
+    boolean::*, column_filter::*, column_filter_ui::*, filter_trait::*, numerical::*,
+    parse_timestamp::*, string::*, timestamp::*, timestamp_formatted::*,
 };
